Move signature validation
